refactored code and project packages
