+++ src/main/java/com/tvlk/payment/ruleengine/core/DefaultRulesEngine.java
@@ -24,6 +24,7 @@
 package com.tvlk.payment.ruleengine.core;
 
 import com.tvlk.payment.ruleengine.Constants;
+import com.tvlk.payment.ruleengine.model.rules.UnitRuleGroupWithResult;
 import org.jeasy.rules.api.Facts;
 import org.jeasy.rules.api.Rule;
 import org.jeasy.rules.api.RuleListener;
